Don't delete output file when debugging is enabled

+++ class.handbrake.php
@@ -649,7 +649,8 @@
 			}
 
 			$arr = file($output_file, FILE_IGNORE_NEW_LINES | FILE_SKIP_EMPTY_LINES);
-			unlink($output_file);
+			if(!$debug)
+				unlink($output_file);
 
 			// Strip out library messages
 			$pattern = "/^(libdvdread|libdvdnav|libbluray)/";
